bring up Led (output) control to foundation part
