+++ RTOS_STM32/bg/Inc/task2.hpp
@@ -15,9 +15,6 @@ namespace Task
       { }
       static void task2Run(void* param);
       void run(void* param) override;
-
-    private:
-      LEDHandler m_LEDHandler{GPIOD, GPIO_PIN_13};
   };
 
   extern Task2Handler task2;
